Move rc/scheduler.go to its own package
Improve error checking and close channels when they are not used anymore
Move scheduler.go to its own package from rc
Fix more references to rc/scheduler
